// src/types/ppro.d.ts
export declare type Action = {};
// ...
export declare type ClipProjectItem = {
  getInputLUTID(): Promise<string>; //Get Guid of Input LUT overridden on media
  createSetInputLUTIDAction(stringLUTID: string): Action; //Create action for setting Guid of Input LUT on media. This applies for Video Clips only.
  isSequence(): Promise<boolean>; //Returns true if the project item is a sequence
  canChangeMediaPath(): Promise<boolean>; //Returns true if Premiere Pro can change the path associated with this project item; otherwise, returns false
  isOffline(): Promise<boolean>; //Returns true if the media is offline
  canProxy(): Promise<boolean>; //Indicates whether it is possible to attach a proxy to this project item.
  getProxyPath(): Promise<string>; //Returns the proxy path if the project item has a proxy attached
  hasProxy(): Promise<boolean>; //Indicates whether a proxy has already been attached to the project item.
  attachProxy(
    mediaPath: string,
    isHiRes: boolean,
    inMakeAlternateLinkInTeamProjects?: boolean,
  ): Promise<boolean>; //Attach proxy or high resolution footage to projectItem and returns true if successful. Not undoable.
  findItemsMatchingMediaPath(
    matchString: string,
    ignoreSubclips?: boolean,
  ): Promise<ProjectItem[]>; //Returns array of project's items with media paths containing match string
  refreshMedia(): Promise<boolean>; //Updates representation of the media associated with the project item
  createSetOfflineAction(): Action; //Returns an action which sets the media offline
  getFootageInterpretation(): Promise<FootageInterpretation>; //Get the footage interpretation object for project item
  createSetFootageInterpretationAction(
    footageInterpretation: FootageInterpretation,
  ): Action; //Set the footage interpretation object for project item
  changeMediaFilePath(
    newPath: string,
    overrideCompatibilityCheck?: boolean,
  ): Promise<boolean>; //Change media file path of projectItem and returns true if successful. Not undoable.
  isMergedClip(): Promise<boolean>; //Returns true if the clip Project item is a merged clip
  isMulticamClip(): Promise<boolean>; //Returns true if the clip Project item is a multicam clip
  getEmbeddedLUTID(): Promise<string>; //Get GUID of LUT embedded in media
  createSetScaleToFrameSizeAction(): Action; //Returns an action which sets the scale to frame to true
  createSetNameAction(inName: string): Action; //Returns action that renames projectItem
  getColorLabelIndex(): Promise<number>; //Get color label index of projectItem
  createSetColorLabelAction(inColorLabelIndex: number): Action; //Create an action for set color label to projectItem by index
  getProject(): Promise<Project>; //Get the parent Project of this projectItem.
  getContentType(): Promise<Constants.ContentType>; //Get content type of the Project item
  getSequence(): Promise<Sequence>; //Get the sequence of the Project item
  getInPoint(mediaType: Constants.MediaType): Promise<TickTime>; //Get the in point of the Project item
  getOutPoint(mediaType: Constants.MediaType): Promise<TickTime>; //Get the out point of the Project item
  getMediaFilePath(): Promise<string>; //Get the media file path of the Project item.
  getComponentChain(mediaType: Constants.MediaType): Promise<string>; //Get the media file path of the Project item.
  createSetInPointAction(tickTime: TickTime): Action; //Returns an action which Sets the in point of the Project item
  createSetOverridePixelAspectRatioAction(
    inNumerator: number,
    inDenominator: number,
  ): Action; //Returns an action which sets Override pixel aspect ratio
  createSetOverrideFrameRateAction(inOverriddenFrameRateValue: number): Action; //Returns an action which sets the override frame rate
  createSetOutPointAction(tickTime: TickTime): Action; //Returns an action which Sets the in point of the Project item
  createSetInOutPointsAction(inPoint: TickTime, outPoint: TickTime): Action; //Set the in or out point of the Project item
  createClearInOutPointsAction(): Action; //Create Clear the in or out point of the Project item action
  getMedia(): Promise<Media>; //Return media associated with clipProjectItem
  type: number; //Get the type of the Project Item.
  name: string; //The name of this project item.
};
// ...
export declare type FolderItem = {
  createBinAction(name: string, makeUnique: boolean): Action; //Returns an action that lets users create a new bin.
  createSmartBinAction(name: string, searchQuery: string): Action; //Creates a smart bin with given name and returns the Folder object
  createRenameBinAction(name: string): Action; //Rename the Bin and return true if it's successful
  getItems(): Promise<ProjectItem[]>; //Collection of child items of this folder.
  createRemoveItemAction(item: ProjectItem): Action; //Creates an action that removes the given item from this folder.
  createMoveItemAction(item: ProjectItem, newParent: FolderItem): Action; //Creates an action that moves the given item to the provided folder item newParent.
  createSetNameAction(inName: string): Action; //Returns action that renames projectItem
  getColorLabelIndex(): Promise<number>; //Get color label index of projectItem
  createSetColorLabelAction(inColorLabelIndex: number): Action; //Create an action for set color label to projectItem by index
  getProject(): Promise<Project>; //Get the parent Project of this projectItem.
  type: number; //Get the type of the Project Item.
  name: string; //The name of this project item.
};
// ...
export declare type Project = {
  getActiveSequence(): Promise<Sequence>; //Get the active sequence of the project
  setActiveSequence(sequence: Sequence): Promise<boolean>; //Set the active sequence of the project
  createSequence(name: string, presetPath?: string): Promise<Sequence>; //Create a new sequence with the default preset path - Parameter presetPath is deprecated, instead use createSequenceWithPresetPath()
  createSequenceFromMedia(
    name: string,
    clipProjectItems?: ClipProjectItem[],
    targetBin?: ProjectItem,
  ): Promise<Sequence>; //Create a new sequence with a given name and medias
  getColorSettings(): Promise<ProjectColorSettings>; //Get project color settings object
  deleteSequence(sequence: Sequence): Promise<boolean>; //Delete a given sequence from the project
  getInsertionBin(): Promise<ProjectItem>; //Get current insertion bin
  openSequence(sequence: Sequence): Promise<boolean>; //Open a sequence and return true if successful.
  importSequences(projectPath: string, sequenceIds?: Guid[]): Promise<boolean>;
  importAEComps(
    aepPath: string,
    compNames: string[],
    TargetBin?: ProjectItem,
  ): Promise<boolean>;
  importAllAEComps(aepPath: string, TargetBin?: ProjectItem): Promise<boolean>;
  importFiles(
    filePaths: string[],
    suppressUI?: boolean,
    targetBin?: ProjectItem,
    asNumberedStills?: boolean,
  ): Promise<boolean>; //Import files in root/target bin of the project
  close(closeProjectOptions?: CloseProjectOptions): Promise<boolean>; //Close a project
  save(): Promise<boolean>; //Save the project
  saveAs(path: string): Promise<boolean>; //Save the project at the provided path
  getSequence(guid: Guid): Sequence; //Get sequence by id from the project
  getSequences(): Promise<Sequence[]>; //Get an array of all sequences in this project.
  getRootItem(): Promise<FolderItem>; //The root item of the project which contains all items of the project on the lowest level.
  pauseGrowing(pause: boolean): Promise<boolean>; //Pause growing of files instead swap the files
  executeTransaction(
    callback: (compoundAction: CompoundAction) => void,
    undoString?: string,
  ): boolean; //Execute undoable transaction by passing compound action
  lockedAccess(callback: () => void): void; //Get a read/upgrade locked access to Project, project state will not change during the execution of callback function. Can call executeTransaction while having locked access.
  guid: Guid; //The unique identifier of the project.
  name: string; //The project name.
  path: string; //The absolute file path to the project file.
};

export declare type ProjectClosedEventStatic = {
  EVENT_CLOSED: string; //Event occurs when project was closed.
};

export declare type ProjectClosedEvent = {
  name: string; //The project name.
  path: string; //The absolute file path to the project file.
  id: string; //The unique identifier of the project.
};

export declare type ProjectColorSettings = {
  getSupportedGraphicsWhiteLuminances(): Promise<number[]>; //Get all the graphics white luminance as array of values
  getGraphicsWhiteLuminance(): Promise<number>; //Get the graphics white luminance value
};
// ...
export declare type ProjectItem = {
  createSetNameAction(inName: string): Action; //Returns action that renames projectItem
  getColorLabelIndex(): Promise<number>; //Get color label index of projectItem
  createSetColorLabelAction(inColorLabelIndex: number): Action; //Create an action for set color label to projectItem by index
  getProject(): Promise<Project>; //Get the parent Project of this projectItem.
  getId(): string; //Get id of projectItem
  getParentBin(): FolderItem; //Get parent FolderItem of projectItem
  type: number; //Get the type of the Project Item.
  name: string; //The name of this project item.
};
// ...
export declare type Sequence = {
  getSequenceVideoTimeDisplayFormat(): Promise<TimeDisplay>; //Get video time display format of this sequence
  getSequenceAudioTimeDisplayFormat(): Promise<TimeDisplay>; //Get audio time display format of this sequence
  getPlayerPosition(): Promise<TickTime>; //Get the player's current position
  setPlayerPosition(positionTime?: TickTime): Promise<boolean>; //Set the player's current position
  clearSelection(): Promise<boolean>; //Clears TrackItem Selection
  setSelection(trackItemSelection: TrackItemSelection): Promise<boolean>; //Updates sequence selection using the given track item selection.
  getVideoTrackCount(): Promise<number>; //Get video track count from this sequence
  getAudioTrackCount(): Promise<number>; //Get audio track count from this sequence
  getCaptionTrackCount(): Promise<number>; //Get caption track count from this sequence
  getVideoTrack(trackIndex: number): Promise<VideoTrack>; //Get video track from track index
  getAudioTrack(trackIndex: number): Promise<AudioTrack>; //Get audio track from track index
  getCaptionTrack(trackIndex: number): Promise<CaptionTrack>; //Get caption track from track index
  getSettings(): Promise<SequenceSettings>; //Get sequence settings object
  createSetSettingsAction(sequenceSettings: SequenceSettings): Action; //Returns action that set sequence settings
  createCloneAction(): Action; //Creates an action to clone the given sequence
  createSubsequence(ignoreTrackTargeting?: boolean): Promise<Sequence>; //Returns a new sequence, which is a sub-sequence of the existing sequence
  isDoneAnalyzingForVideoEffects(): Promise<boolean>; //Returns whether or not the sequence is done analyzing for video effects
  getZeroPoint(): Promise<TickTime>; //Time representing the zero point of the sequence.
  getEndTime(): Promise<TickTime>; //Time representing the end of the sequence
  getInPoint(): Promise<TickTime>; //Get time representing the inPoint of sequence.
  getOutPoint(): Promise<TickTime>; //Get time representing the inPoint of sequence.
  createSetInPointAction(tickTime: TickTime): Action; //Create SetInPointAction for sequence
  createSetZeroPointAction(tickTime: TickTime): Action; //Create an action to set an InPoint for the sequence
  createSetOutPointAction(tickTime: TickTime): Action; //Create SetOutPointAction for sequence
  getProjectItem(): Promise<ProjectItem>; //Get the associated projectItem of the sequence.
  getSelection(): Promise<TrackItemSelection>; //Returns the current selection group of the sequence.
  getFrameSize(): Promise<RectF>; //Gets the size of the frame
  getTimebase(): Promise<string>; //Gets the time base of sequence
  guid: Guid; //The unique identifier of the sequence.
  name: string; //The sequence name.
};
// ...
  export enum ContentType {
    ANY,
    SEQUENCE,
    MEDIA,
  }